refactor(account-management-service): In VerifyService, the methods have been redesigned to extract the username from the token

+++ email-verify-parent/email-verify-service/src/main/java/org/capsule/com/services/VerifyService.java
@@ -1,11 +1,13 @@
 package org.capsule.com.services;
 
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.capsule.com.configs.Constants;
 import org.capsule.com.configs.Message;
+import org.capsule.com.dtos.errors.SomeErrorMessage;
 import org.capsule.com.dtos.errors.WrongField;
 import org.capsule.com.dtos.requests.CodeConfirmReqst;
 import org.capsule.com.dtos.requests.UserInfoReqst;
@@ -14,6 +16,8 @@ import org.capsule.com.services.producers.KafkaJsonProducerService;
 import org.capsule.com.services.producers.KafkaStringProducerService;
 import org.capsule.com.utils.exceptions.NotValidException;
 import org.capsule.com.utils.tools.CodeGeneratorTool;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -23,8 +27,10 @@ import org.springframework.validation.BindingResult;
 @RequiredArgsConstructor
 public class VerifyService {
 
+    private final Logger LOGGER = LoggerFactory.getLogger(VerifyService.class);
     private final VerifyDBService verifyDBService;
     private final CodeGeneratorTool codeGeneratorTool;
+    private final DecodeJWTService decodeJWTService;
     private final KafkaJsonProducerService kafkaJsonProducerService;
     private final KafkaStringProducerService kafkaStringProducerService;
 
@@ -32,9 +38,14 @@ public class VerifyService {
         BindingResult bindingResult) {
         validate(bindingResult);
 
-        Verify verify = createEntity(info);
+        String username = extractUsernameFromToken(token);
+        Verify verify = createEntity(username, info.getEmail());
         verifyDBService.save(verify);
         kafkaJsonProducerService.produce(Constants.LETTERS_WITH_CODE_TOPIC, verify);
+
+        LOGGER.info(
+            "VerifyService [email-verify-service] the verification request has been accepted for [{}] to [{}]",
+            username, info.getEmail());
 
         return ResponseEntity.ok(HttpStatus.OK);
     }
@@ -43,20 +54,30 @@ public class VerifyService {
         BindingResult bindingResult) {
         validate(bindingResult);
 
-        String username = code.getUsername();
+        String username = extractUsernameFromToken(token);
         Verify verify = verifyDBService.findByUsername(username);
-        if (verify.getCode() == code.getCode()) {
-            verifyDBService.deleteAllByUsername(username);
+        if (verify.getCode() == code.getCode()
+            && isWithinTwoMinutes(LocalDateTime.now(), verify.getCreatedAt())) {
             kafkaStringProducerService.produce(Constants.SUBMIT_VERIFY_STATUS_TOPIC, username);
+            verifyDBService.deleteAllByUsername(username);
+            LOGGER.info("VerifyService [email-verify-service] successful verification for [{}]", username);
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            LOGGER.info("VerifyService [email-verify-service] failed verification for [{}]", username);
+            throw new NotValidException("the code is already invalid",
+                List.of(new SomeErrorMessage("the code is already invalid")));
         }
     }
 
-    private Verify createEntity(UserInfoReqst info) {
-        String username = info.getUsername();
-        String email = info.getEmail();
+    private String extractUsernameFromToken(String token) {
+        try {
+            return decodeJWTService.extractUsername(token.substring(7));
+        } catch (Exception ex) {
+            throw new RuntimeException(ex.getMessage());
+        }
+    }
+
+    private Verify createEntity(String username, String email) {
         int code = codeGeneratorTool.generate();
         LocalDateTime createdAt = LocalDateTime.now();
         return Verify.builder()
@@ -79,5 +100,10 @@ public class VerifyService {
 
             throw new NotValidException(Message.NOT_VALID_FIELDS_MSG, wrongFields);
         }
+    }
+
+    private boolean isWithinTwoMinutes(LocalDateTime first, LocalDateTime second) {
+        Duration duration = Duration.between(first, second);
+        return Math.abs(duration.toSeconds()) < 120;
     }
 }
